requirements.txt
flask
gunicorn
pandas
plotly
scikit-learn
fpdf
reportlab
matplotlib
statsmodels
python-dotenv
thefuzz
tqdm
kiwipiepy
openai